refactor(sandbox): decouple generic core from the Python profile
The generic Profile-A core (Sandbox) no longer knows about Python. Add a generic
Sandbox::call_export(name, input, fuel) -> i32 that ships a byte payload to a named
guest export over the guest alloc/memory ABI; the shared run_export body is reused by
both the fuel- and epoch-bounded paths.

Move the Python-specific surface into a new `python` profile module:
  * Sandbox::eval / eval_with_epoch_deadline (hardcoded "eval" export) — now built
    on call_export / run_export
  * open_shell, PyShell, PyOp, PyResult
  * the /lang/python mount (PythonMount) -> src/python/mount.rs

Profile B (WasiSandbox / wasi_fs) is already generic and untouched. The crate root
doc now reads as a generic wasm-guest sandbox host with python as one profile.

Also rewrite spike-era doc prose ("The whole point", "vs the P0 spike", #483/#488
P1/P2 references) as present-tense API docs.

Drop dead helper as_mount from the mount module.
